refactor(040): migra apps/web/src/features/medications para TypeScript
