+++ apps/web/src/features/medications/components/LaboratoryAutocomplete.tsx
