+++ apps/web/src/features/medications/services/__tests__/medicineDatabaseService.test.ts
@@ -6,10 +6,11 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
  * para devolver `load()` determinístico (sem fetch/JSON). O fetch/cache/offline/timeout
  * reais são cobertos no teste do core (anvisaDatabase.test.js).
  */
-let mockData = []
+// eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO(040-strict) tipar
+let mockData: any[] = []
 
 vi.mock('@dosiq/core', async (importActual) => {
-  const actual = await importActual()
+  const actual = (await importActual()) as any
   return {
     ...actual,
     createAnvisaDatabase: () => ({ load: () => Promise.resolve(mockData) }),
